AudioRobot_task_for_Deepgram/speech_to_text/requirements.txt
pyaudio
python-dotenv
deepgram-sdk==3.*

langchain
langchain_community
langchain-openai
langchain-groq